Check object and object[] on union type and after fixes
The autofix from Object[] to object[] was done after the check for
object[] which shows this up on the second run and break libup.
Libup build the exclude list from the first run and than autofix.
Everything coming up after autofix is a problem for libup.
Just move the line to avoid further problems

Change-Id: I5687a7cd80527102b3e0c60b401b78d6e4a8dcb1