Avoid slow strcasecmp() where not necessary
While it is possible to write functions like DIRNAME(),
ARRAY_KEYS() and such in upper-case or mixed-case, this is so
unlikely, I feel it is not worth checking for such extreme
edge-cases. This small changes cuts the runtime of the
DirUsage sniff in half (!), and speeds up the InArrayUsage
sniff as well.

While these two sniffs are certainly not the slowest ones we
have, their runtime is still a notable fraction of the total
runtime.

Change-Id: I686a6a21dec61991e1bdc28830cd4093894f1026