Rewrite EmptyTag sniff for performance
When I benchmark an example run on a medium size codebase, this
sniff is the 7th slowest one. I was wondering why. It doesn't
do that much. It shouldn't be that slow.

This rewrite cuts the runtime of this sniff down to effectively
zero.

Relative runtime of this sniff:
* 0.55% before
* 0.02% after

Number of times the sniff is called:
* 9300 times before
* 800 times after

The main reason for this difference is because the old
implementation was called for every variable token. There are
thousands and thousands of variables in a codebase, but only
very few do have a PHPDoc comment. The new implementation
turns the two logical things the code does around:
1. Look for empty PHPDoc tags.
2. Only if we found one (which shouldn't happen in a clean
   codebase) check if the PHPDoc block belongs to a function
   or class property.

Change-Id: I06d94c43fd3fb7908e47cb626daf2f146d651def