Rewrite ParenthesesAroundKeywordSniff to fix whitespace handling
The use of two arrays for next and nextnext was confusing when checking
for whitespaces.
Now search for the next non-whitespace token, when it is a ( and has a
closer process furter.
Than add a whitespace if the keyword and ( are together.
Remove the ( and an optional whitespace, after that remove ) and the
whitespace before it.

Adding a whitespace and removing the whitespace after the opener is
easier to follow than having code checking if the whitespace after the
opener may be kept or still need a new one

Bug: T265741
Change-Id: Ic2a9bec74df5ba0c85118c8b483c5a2fb0641f12