Fix performance bottleneck in PHPUnit assert sniffs
The performance impact for the two AssertCount and AssertEquals
sniffs is massive. In a test run on a medium size codebase
these two sniffs are triggered 12,000 times before the change
(= total number of string tokens in the codebase), but 160
times after (= number of files in the codebase). The runtime is
a significant percentage of the total runtime before the change,
but effectively zero after (about 250x faster).

The two DeprecatedPHPUnitMethods and PHPUnitTypeHints sniffs
aren't triggered for string tokens but for class tokens. There
is nothing to improve. Still I feel it's good practice to make
the code behave the same.

Change-Id: I40979bf2e18f75772e67bbc03a5dbb38f4945521