Cache PHPUnitTestTrait::isTestFile() results
This is currently the by far slowest piece of code in this
codebase. It's used in multiple sniffs we don't want to run
outside of tests.

In case a file is not a test file, it's skipped. In other
words, the performance of this method isn't a problem in a
codebase without tests. isTestFile() is executed once per
file, and that's it.

But if it's a test, this method is executed thousands and
thousands of times for every single string in the file. This
piles up. With a caching layer in place the expensive code is
executed once per file. This improves the performance of the
affected sniffs by a factor of 10.

Another option to solve this issue is to rearrange the
conditionals in all sniffs to be ordered from the cheapest to
the most expensive ones. But this is brittle, hard to
understand, not as effective, and will be forgotten when
working on future sniffs.

Change-Id: Id6d6dd34c5dfdac9d52cc0a029b50252f2ab0f07