Micro-optimization for UnusedUseStatementSniff
Quick local tests suggest that the runtime:
 - Goes from 46 to 23 for `"bool[] $var Comment"`
 - Goes from 76 to 36 for `$var bool[] Comment`
 - Goes from 183 to 84 for `'array<int, array<string, bool>> $var
   Comment'`

These results were obtained with 1m repetitions in an outer loop of 10
repetitions, taking the olympic mean from the outer loop. The
benchmarking scripts uses hrtime (on PHP 7.4.1), which makes the results
very consistent, with a SD < 1. Times above in ms.

Not sure if it's going to make any difference in a real-world run, but
it's probably worth doing anyway.

Change-Id: Id2295ddaeea44e751a37d361f7ab009dbdb8faa0