Improve performance of exclude-pattern in ruleset.xml
When the exclude-pattern ends with "/*" phpcs applies the pattern to
folders and avoids recursion into the subfolder.
That means for the vendor folder not each file is checked, only the
folder name once.
For this repo this cut down from around 9000 files to 114 in the
internal processing, while the sum of checked files is the same.
It saves 2 second runtime on this repo.

Exclude /coverage/ folder as well to avoid scanning the subfolder
Only holds derived content and it is known to have no php files

Combine all into one exclude-pattern to avoid a loop over each simple
regex. That possible only a micro optimisation.

Change-Id: Id95236edbd3d87f57719d235c7c9cee366180099