Forbid eval() and make ForbiddenFunctions sniff more flexible
eval() is usually considered – well – evil. While it has it's use in
places, it's remarkably rare:
https://codesearch.wmcloud.org/search/?q=%5C%20%5Cbeval%5C(&files=%5C.php%24

I would like to argue that it is beneficial when devs are forced to
think about it and either mark these lines of code with a "phpcs:ignore"
or document the function as being allowed in .phpcs.xml.

This patch also improves the sniff:
* Do the check for the function brackets as early as possible.
* Make it possible to add custom entries as well as modify existing
  ones.
* Make it possible to use the string "null". This is necessary for the
  <property> in .phpcs.xml.
* Make it possible to use no-op replacements like "foo=>foo".

Change-Id: Iaf089e7a07d5b4000939be78b73e8b03db885e0c