Rewrite internal PHPUnit test runner for speed
In my local test this reduces the total runtime for all PHPUnit tests
in this codebase from 8s to 5s. This is because we did everything twice
before. One time to get the report (what PHPCS would output), and
another time for the fixes (what PHPCBF would output). We can as well
do both together.

Change-Id: Icaa1ff636c4bf74168ddce4441578e6788e05cac