Allow typed properties to omit documentation
Properties are allowed to omit their documentation comment if they have
a type declaration. Also, fix the variable name of $methodProps /
$memberProps (copy+paste?).

In the test, put the `var` line first, before the ones with proper
visibility (`public` etc.), unlike in the previous properties, so that
this is line 97; if it was at the end, it would be line 100, and the
entire .expect file would change because the first column (with the line
numbers) would now be one character wider.

Bug: T288754
Change-Id: I446568e397685f55afb82e3819c451b56e6cbcbd