Allow prose-only documentation for statically typed functions
Since change I4f797fa8c4, functions that have static parameter types and
a static return type are allowed to have no documentation comment at
all; additionally, we want to allow them to have documentation comments
that contain prose (and potentially @throws etc.), but neither @param
nor @return tags. Also, tests are still allowed to not document any
params, as long as they do it consistently.

If at least one parameter is documented (e.g. because its type can’t be
represented in the PHP versions we currently support), then we still
demand that *all* parameters are documented, even if for some of them
the phpdoc type is redundant with the static type; I think allowing the
phpdoc to only list a subset of parameters would be too confusing.

Bug: T258925
Change-Id: Ie1be29a10f777ca73e9ba73c77d638837d0af675