Enable Generic.WhiteSpace.ScopeIndent
This sniff ensures that all code inside a scope whose opening has an
indentation of N tabs will have at least N+1 tabs of indentation.

Note that this sniff does not complain about code that is indented too
far.

Some tokens can legitimately be placed at the beginning of the line or
otherwise be indented differently than their siblings within the same
scope. Those that are found in MediaWiki core were excluded in the ruleset.

Most of the remaining errors in MW core are non-indented `switch`es and
messy array indentation (especially when the array is an argument inside
a multiline function call). There are also a couple methods with wrong
indentation, and the whole ZhConversion class lacks indentation.

Bug: T334996
Change-Id: I1e85a43e3f5626ca017893f25553d6a9fbcaaaf0