Enable Squiz.ControlStructures.ControlSignature
This sniff verifies that spacing around control structures is correct.
It supersedes the following custom sniffs, that were thus removed:
- SpaceAfterControlStructureSniff
- IfElseStructureSniff
- SpaceBeforeControlStructureBraceSniff
- SameLineCatchSniff

On top of that, it can validate the following scenarios that previous
sniffs would not catch:
- T_DO, T_ELSE, T_FINALLY, T_MATCH and T_TRY tokens are handled
  consistently in the same place as the others
- Enforce spacing after closing braces
- Enforce spacing between closing parens and opening brace
- Enforce 0 spaces after IF and ELSEIF when using alternative syntax
  (the previous sniffs would only handle ELSE)

All tests for the previous sniffs were consolidated in a single file, to
make sure that the upstream sniff is enforcing the style we want.

Bug: T334993
Bug: T311282
Change-Id: I2c045c70125b64af25ca684eaa68fddecc556e10