Actually check if a method returns anything
As of now the sniff sometimes (!) complains about a missing return
type documentation without checking if the method actually returns
anything. This was behaving inconsistent, depending on the code path
taken.

Issue found in Ic82e8b2.

This patch introduces a third class: We can't know if a method
returns anything when it's part of an interface or abstract. In this
case we …
1. … force users to add either a PHPDoc comment or an actual return
   type, e.g. `: void`.
2. … but assume everything is fine when there is some comment, no
   matter what the comment says.

The later might change in the future, i.e. we might want to enforce
`: void` everywhere. But we are not there yet and it's not the
intention of this patch to change that.

Change-Id: Iae9b9bb635575b3925f9f8dd9dfbc31c04ee029a