Enable the `Universal.WhiteSpace.CommaSpacing` sniff
This sniff enforces the correct spacing around commas (no space before,
and exactly one space or a newline after). The sniff provides an
autofixer.

It currently emits 148 warnings in MW core, and most of them seem to be
correct. The main exceptions are in-line comments where the author
wanted to achieve vertical alignment of comments, like this:

  [
	'foo',    # Some comment here
	'foobar', # Another comment here
	'x',      # And another here
  ]

This mostly occurs in PHPUnit data providers. However:
- Our coding conventions explicitly forbid vertical alignment: https://www.mediawiki.org/wiki/Manual:Coding_conventions#Vertical_alignment
- The MediaWiki.WhiteSpace.SpaceBeforeSingleLineComment.NewLineComment,
  enabled by default, forbids inline comments (it's disabled in lots of
  places, though)
- In some of the examples I found (~1/3), vertical alignment is already
  broken, I guess because somebody changed some of the lines without
  re-aligning everything (vertical alignment is forbidden for a
  reason...).

Therefore, I would considere the code in question to be non-compliant,
and flagging it seems the right thing to do.

This new sniffs generalizes the existing OneSpaceInlineArray, which can
therefore be removed.

Bug: T349327
Bug: T347438
Change-Id: Iccd11bfae43990228a803ef75d40544edd53bfeb