Replace AlternativeSyntaxSniff with upstream sniff
It does the same thing, but it is more thorough in that it actually
reports an issue on every line where the alternative syntax is used (not
just the closing keyword), and it offers an autofixer.

Retain the existing test to make sure that the MW standard as a whole
forbids alternative syntax.

Bug: T349327
Change-Id: I404cf2ebfc30c71e6c3d3ad7d40265828901abf7