Add test for awfully inconsistent code
I've asked ChatGPT to generate a snippet of highly inconsistent PHP
code. Then I made it even worse. This should give us a test with some
extreme edge cases to see what our PHPCS standards do or do not catch.

For this test, the .fixed result is particularly interesting, as it
shows what PHPCS would consider correct after a pass of phpcbf. This can
be compared to what a human would deem to be correctly-formatted code.

Change-Id: Ifd4661bf5e722aa500140b6a4651481dc342116a