Replace DoubleNotOperatorSniff with Universal.CodeAnalysis.NoDoubleNegative
The two sniffs are roughly equivalent. The upstream one has an
additional error messages for when >= 3 NOT operators are found, and the
error message for the instanceof case is more specific, too. Also, our
own sniff checked for `instanceof` anywhere in the expression, hence
refusing to fix `!!$x || $a instanceof B`, which is actually safe to
fix. The upstream sniff autofixes that.

Add a single testcase to generic_fail to document that we disallow
double negatives in our coding standards, and remove the existing
dedicated tests: let the edge cases be tested upstream.

Change-Id: I1c15cae499d6015dcc6974db18814298e8468eba