SpaceyParenthesisSniff: detect unnecessary whitespace in more cases
When checking for unnecessary whitespace within parentheses, look for
the next non-whitespace token before the closer parenthesis, rather than
trying to count the tokens. This allows us to catch more cases of
unnecessary whitespace, as demonstrated by the new tests.

For arrays (short or long), preserve the status quo and do not complain
when there's a newline between the brackets. In MW core alone there are
quite a few examples of empty arrays with a newline inside.

Also, change return statements to consistently use the position of the
closer parenthesis, so it's easier to see what is being done.

Add a few test cases, and disable the rule for long arrays as it's not
relevant to this sniff.

Change-Id: Id8a9bb037a485574a8423540d049e118ccb7598c