Add tests for closure declarations
Currently, our ruleset is very permissive with closure declarations, and
does not flag code that would normally be flagged for normal
(non-closure) declarations. Turns out the
OpeningFunctionBraceKernighanRitchieSniff sniff has a configurable
`checkClosures` property, false by default and that we never enabled.

The added tests documents the unwanted behaviour. A subsequent patch
will enable the closure checks, to show the differences.

Change-Id: Ic9a9f82c5ac4708982e51b48232cc56606821b92