EmptyLinesBetweenUseSniff: fix edge case with closures
The sniff in question never checked that it was actually dealing with
namespace `use`. This led to an edge case under the following
conditions:
- The first `use` statement in the file belongs to a closure.
- There is at least another `use` after that.
- The second `use ` statement is the first non-empty token on its line.

This is quite unlikely to happen in the real world, but it could be
triggered in tests: for example, see PS 4 of Ic9a9f82c5ac47.

OTOH, checking for closure `use` is cheap and only needs to be done once
per per file, so just do that.

Change-Id: Id99b1f35234e9b85481c078dc8cb83203a800b8c