Introduce CommentAnnotationsTrait to deduplicate some code for doc tags
This lets us share more code between FunctionAnnotationsSniff and
PropertyAnnotationsSniff; and in a subsequent patches, a similar sniff
for classes.

Also bring the lists of allowed annotations all in the same place, and
deduplicate them by keeping a list of annotations allowed everywhere.

The lists themselves are unchanged, with the exception of
`@phan-read-only` and `@phan-write-only` now being allowed on
properties, since the former is needed for this very patch.

Change-Id: Ifd8db9b4c94b00bcc388d7d685ccfeb9244bc7d8