Allow the `@template` annotation
In I044ef8847d12ae, we went with the more specific version because I
wasn't sure if there were any tools outside phan for which the
`@template` annotations worked differently.

Today it seems clear that all tools that support `@template` support
roughly the same syntax, so it makes sense to allow the unprefixed
version.

Bug: T397781
Change-Id: I22873f2b5bf3398843d29935039ef6099d597838